+++ test/unit/rmq_metadata/monitor_queue.py
@@ -83,8 +83,6 @@ class CfgTest(object):
              "postname": "",
              "mode": "w",
              "ext": "pdf",
-             "dtg": False,
-             "date": False,
              "stype": "encoded",
              "archive": False}]
         self.mongo_cfg = "mongo"