+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 ## [0.0.2] - 2020-09-29
 ### Changed
 - config/rabbitmq.py.TEMPLATE:  Removed dtg and date entries from queue_list.
+- Documentation updates.
 
 
 ## [0.0.1] - 2020-09-11